+++ dev_lib/system/delay.h
@@ -4,6 +4,6 @@
 #include "stm32f4xx_hal.h"
 
 void delay_us(uint32_t us);
-uint32_t delay_get_us_tick(void);
+uint64_t delay_get_us_tick(void);
 
 #endif
